requirements.txt
django
djangorestframework
httpie

